refactor library index files
