+++ packages/layers/rollup.config.mjs
@@ -2,6 +2,6 @@ import config from '../../rollup.config.mjs';
 
 export default {
   ...config,
-  input: './src/index.tsx',
+  input: './src/index.ts',
   plugins: [...config.plugins]
 };
