+++ packages/layers/src/layers/index.ts
@@ -2,3 +2,4 @@ export * from './LayerContext';
 export * from './useLayer';
 export * from './LayerNode';
 export * from './DefaultLayer';
+export * from './Layers';
