+++ packages/layers/tsconfig.json
@@ -4,5 +4,5 @@
     "baseUrl": "./src",
     "outDir": "./dist"
   },
-  "include": ["src/index.tsx"]
+  "include": ["src/index.ts"]
 }
